refactor reporting batch retry route
